Refactors app architecture and UI for maintainability
Improves code maintainability and separation of concerns by refactoring the application architecture.

Introduces a controller class to manage application state and business logic, separating it from the UI components.

Updates the UI to use a more modular structure and modern styling with ttk themes.

Adds documentation for the new architecture.
