+++ SafePDF/ARCHITECTURE.md
@@ -1,0 +1,112 @@
+# SafePDF - Refactored Architecture
+
+## Project Structure
+
+The SafePDF application has been refactored to separate concerns between UI and business logic:
+
+### Files and Directories:
+
+```
+SafePDF/
+├── safe_pdf_app.py          # Main application coordinator
+├── safe_pdf_controller.py   # Business logic and state management
+├── pdf_operations.py        # PDF processing operations
+├── ui/
+│   ├── __init__.py         # UI package initialization
+│   └── safe_pdf_ui.py      # All UI components and styling
+├── assets/                 # Image assets for operation buttons
+├── welcome_content.html    # Welcome tab HTML content
+├── welcome_content.txt     # Welcome tab text fallback
+└── requirements.txt        # Python dependencies
+```
+
+## Architecture Overview
+
+### 1. `SafePDFApp` (safe_pdf_app.py)
+- **Role**: Main application coordinator
+- **Responsibility**: 
+  - Initialize the application
+  - Create and coordinate between controller and UI
+  - Handle progress callbacks
+- **Size**: Minimal - acts as a thin coordinator layer
+
+### 2. `SafePDFController` (safe_pdf_controller.py)
+- **Role**: Business logic and state management
+- **Responsibilities**:
+  - File selection and validation
+  - Operation selection and management
+  - Settings management
+  - PDF operations coordination
+  - Asynchronous operation execution
+  - State management (selected file, operation, settings, etc.)
+- **Key Features**:
+  - Thread-safe operation execution
+  - Callback system for UI updates
+  - Centralized application state
+  - Input validation and error handling
+
+### 3. `SafePDFUI` (ui/safe_pdf_ui.py)
+- **Role**: User interface management
+- **Responsibilities**:
+  - Window setup and styling
+  - Tab creation and management
+  - Widget creation and layout
+  - Event handling (drag/drop, clicks, etc.)
+  - Settings UI generation
+  - Progress visualization
+  - User feedback (messages, dialogs)
+- **Key Features**:
+  - Modern UI styling with ttk themes
+  - Drag and drop support
+  - Dynamic settings based on operation
+  - Image-based operation buttons
+  - Responsive layout
+
+## Benefits of This Architecture
+
+### 1. **Separation of Concerns**
+- UI code is completely separate from business logic
+- Controller handles all state management and operations
+- Each component has a clear, single responsibility
+
+### 2. **Maintainability**
+- Easy to modify UI without affecting business logic
+- Business logic changes don't impact UI code
+- Clear interfaces between components
+
+### 3. **Testability**
+- Controller can be tested independently of UI
+- Business logic is isolated and testable
+- Mock UI can be created for testing
+
+### 4. **Extensibility**
+- New operations can be added by extending the controller
+- UI can be enhanced without touching business logic
+- Alternative UIs (CLI, web) could use the same controller
+
+### 5. **Code Organization**
+- Related functionality is grouped together
+- Smaller, focused files are easier to navigate
+- Clear dependency structure
+
+## Usage
+
+The refactored application maintains the same user experience while providing a cleaner, more maintainable codebase:
+
+```python
+# Run the application
+python safe_pdf_app.py
+```
+
+The application will:
+1. Initialize the controller with PDF operations support
+2. Create the UI with modern styling and components
+3. Set up communication between UI and controller
+4. Launch the main application window
+
+## Development Notes
+
+- The `SafePDFController` uses callbacks to communicate with the UI
+- All UI state is managed through the controller
+- Threading is handled safely within the controller
+- The UI focuses purely on presentation and user interaction
