+++ SafePDF/ui/__init__.py
@@ -1,0 +1,10 @@
+"""
+SafePDF UI Package
+v1.0.0 by mcagriaksoy - 2025
+
+This package contains all UI-related components for the SafePDF application.
+"""
+
+from .safe_pdf_ui import SafePDFUI
+
+__all__ = ['SafePDFUI']
